Improve vjp rules for composite functions that are treated as primitives (PR347)
* Return None as grad for unused variables

* Use tree_map read for correct handling of tuples of Variables

* Remove wrap_arg from vjp_symbol_mapper

* Test torch function calls are handled correctly with @thunder_backward

* Fix decomposed_fn_aug_fwd_rule

Previously for a function that appears as a separate symbol on the trace
we used the non-decomposed version during the forward pass and
augmented forward + backward during the backward pass amounting to
additional unnecessary work. Now it's fixed. The effect is for vjp
we're always decomposing if we need to. To avoid decomposition
developer must register a special rule for the composite op.