Disable additional transforms for the PyTorch Autograd path (#74)
Applying transforms on the trace after "forward backward split + transform_for_execution" is not a well-tested path and breaks the careful sorting of operations introduced inside the split_forward_backward function. In the future, we need to revisit the order of transformations and move execution transforms out of the split_forward_backward function. Today this PR is needed to restore distributed communication overlap with computation and unblock fixing TransformerEngine+DataParallel.