add an xfail test for the aliasing problem.
Since we're incorrectly returning an alias, modifications to the
returned tensor will show up in the original tensor. Add a test
that demonstrates this, which should remind us to revisit this
implementation once we properly implement functionalization.

Thanks Luca Antiga for the idea.

Also, it turns out PyTorch fixed an issue with their sparse_coo in
the interim and so that test started passing. So this also removes
the sparse test's xfail.