Use torch langctx as the default one
There are places in the code that set the torch langctx as the default
but they don't have the same effect on the whole thunder import.
Developer users might get confused the first time they use Thunder what
is a langctx and why is it necessary. Currently, Thunder is
PyTorch-focused and let's use the Torch language as the default one.