Add a bunch of benchmarks from NeMo's NeVA
This takes a number of the graphs from our "default" configuration
of NeVA and copies them into benchmarks. This should mean most
people do not have to get NeVA up and running to get things
working.

The benchmarks are not perfect:

* There is a bug in the extraction logic that means we miss some
  graphs.
* Some graphs do not run under thunder at all; they are not
  included here. The common case for this happening is an
	unsupported communication op being buried in the graph.
* The extraction logic tends to dump graphs that are identical sans
  `requires_grad` flipping on some tensors. I only included 1 when
	this happens.

Thanks to Mike Ruberry for the idea on how to do this and Ivan
Yashchuk for technical guidance.