Auto-registered ops: Add ops returning views to `_syms_returning_views` (#1059)
As a follow-up to https://github.com/Lightning-AI/lightning-thunder/issues/957 (https://github.com/Lightning-AI/lightning-thunder/issues/957#issuecomment-2293521400)

When auto-registration is added there's no in-place support yet. As mentioned in #957, the operators that return views can interfere with in-place ops. This PR adds the relative auto-registered symbols to `_syms_returning_views` and `_syms_returning_runtime_dependently_views` and lets the in-place pass deal with it.