Reduce `test_thunderfx_mistral_nemo_small` model size (#1701)
Reducing the models size to speedup CI and make it work in environments with constrained memory size.

I think this change should be fine since with this test we are verifying functionality more than memory footprint. Let me know what you think. 

With this change the peak memory of the test is ~2.6GB instead of ~14GB