Remove requires_grad specification from meta functions (#1769)
Meta functions in Thunder should be straightforward and standalone for developers to implement them easily.

Some meta functions try to specify the `requires_grad` attribute for output TensorProxies, but this attribute is never used in Thunder, and the correct propagation of it is not tested.

The changes are needed to make the diff in https://github.com/Lightning-AI/lightning-thunder/pull/1599 cleaner. That PR tries to remove the `requires_grad` attribute from TensorProxies.

Let's discuss bringing back (and testing) of this attribute propagation in https://github.com/Lightning-AI/lightning-thunder/issues/1768.