Skip `prims.check_tensor_shape_and_metadata` when thunder is torch.compile backend
When thunder is a torch.compile backend and example_inputs have no
ambiguity, i.e. no `torch.SymInt`s in `example_inputs`, then I think we
should comfortably skip the prim in prologue. I expect this change to
help reduce the overhead especially when the number of inputs to thunder
function is large.

Signed-off-by: Masaki Kozuki <mkozuki@nvidia.com>