Adds save_thunderfx_repros to save scripts for all the subgraphs and optionally save fusion region and traces (#2232)
While exploring the split reasons and observing changes in traces and fusion areas when adding missing fallback operators, I realized that having an API to save all repro scripts, with options to also save fusion graphs or traces to files, would be helpful. Before this PR, the user had to do the iteration and save manually.